refactor: inputInfoMsg constants 파일로 분리
