+++ src/pages/createGroup/constants/inputInfoMsg.ts
@@ -1,0 +1,8 @@
+export const InputInfoMsg = {
+  groupNameLength: '10자 이내로 작성해주세요.',
+  groupNameNotAvailable: '이미 사용중인 모임명입니다.',
+  groupNameNotCheck: '중복확인을 해주세요.',
+  groupNameAvailable: '사용 가능한 모임명입니다.',
+  groupNameEmpty: '글모임 이름을 입력해주세요.',
+  emptyText: '',
+};
